Update Custom_libraries.md
Rewrote the latter section of it and expanded it a lot. It's gradually coming together, but I feel like the tone of voice shifts throughout it. Is this a case study offering advice on doing this yourself, or is it documenting what I did? Will need further edits after completion.